% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_parameters.R
\name{change_snp_numeric_type}
\alias{change_snp_numeric_type}
\title{Change numeric type of sNp file output}
\usage{
change_snp_numeric_type(data, format)
}
\arguments{
\item{data}{Dataframe to convert}

\item{format}{Target numeric format (valid options = MA, DB, RI)}
}
\description{
Changes the provided dataframe output to the specified type
}
\examples{
sparam <- data.frame(Frequency = 1:3, dB = -1:-3, Ang = 40:42)
rftk:::change_snp_numeric_type(sparam, "MA")

# # A tibble: 3 x 3
#   Frequency   Mag   Ang
#       <int> <dbl> <int>
# 1         1 0.891    40
# 2         2 0.794    41
# 3         3 0.708    42
}
\keyword{internal}
