% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_interpolation.R
\name{complex_approx}
\alias{complex_approx}
\title{Complex Interpolation}
\usage{
complex_approx(x, y, xout, ...)
}
\arguments{
\item{x}{Numeric vector of the x-coordinates to be interpolated}

\item{y}{Complex vector of the y-coordinates to be interpolated}

\item{xout}{Optional set of numeric values specifying where interpolation is to take place}

\item{...}{Other parameters to be sent to \code{approx}}
}
\description{
Enable \code{approx} to interpolate complex numbers
}
\examples{
complex_approx(1:11, -5:5 + -5:5 * -1i, n = 5)

}
