% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_smith.R
\name{coord_smith}
\alias{coord_smith}
\title{Smith Coordinates}
\usage{
coord_smith()
}
\description{
Smith Coordinates
}
\details{
Plots S-parameter data on a smith chart.  Input data currently needs to be in magnitude/angle (rad).  No scales are provided, data is scaled to z0.
}
\examples{
if(require(ggplot2)) {

 ggplot(dipole, aes(x = ang, y = mag)) +
   geom_point() +
   coord_smith()
   
}
}
\seealso{
\code{\link[=geom_smith]{geom_smith()}}
}
