% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dipole}
\alias{dipole}
\title{Dipole S-parameters}
\format{
A data frame with 2001 rows and 3 variables:
\describe{
\item{frequency}{Frequency, in Hz}
\item{mag}{Return \emph{gain} magnitude, linear}
\item{ang}{Return \emph{gain} angle, in degrees}
}
}
\usage{
dipole
}
\description{
A dataset containing the S-parameters of a single dipole
}
\keyword{datasets}
