% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_parameters.R
\name{get_column_names}
\alias{get_column_names}
\title{S-Parameter Column Names}
\usage{
get_column_names(parameter = "S", numeric_format, num_parameters)
}
\arguments{
\item{parameter}{Type of parameter file (valid options = S, Y, Z)}

\item{numeric_format}{Numeric format used (valid options = MA, DB, RI)}

\item{num_parameters}{Number of parameters}
}
\description{
Outputs the columns names of a .sNp file based on parameter, numeric format,
and the number of S-parameters.
}
\examples{
rftk:::get_column_names("S", "MA", 2)
# output: "Frequency" "S11_Mag"   "S11_Ang"   "S21_Mag"   "S21_Ang"   "S12_Mag"   "S12_Ang"
#  "S22_Mag"   "S22_Ang"
}
\keyword{internal}
