% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{impedance_conversion}
\alias{impedance_conversion}
\alias{gamma_to_mismatch}
\alias{gamma_to_s11}
\alias{gamma_to_z}
\alias{mismatch_to_gamma}
\alias{s11_to_gamma}
\alias{z_to_gamma}
\alias{mismatch_to_s11}
\alias{s11_to_mismatch}
\alias{z_to_s11}
\alias{z_to_mismatch}
\title{Impedance Conversion}
\usage{
gamma_to_mismatch(...)

gamma_to_s11(...)

gamma_to_z(..., z0 = 50)

mismatch_to_gamma(...)

s11_to_gamma(...)

z_to_gamma(..., z0 = 50)

mismatch_to_s11(...)

s11_to_mismatch(...)

z_to_s11(..., z0 = 50)

z_to_mismatch(..., z0 = 50)
}
\arguments{
\item{...}{Input reflection coefficient, return gain, impedance, or mismatch
gain}

\item{z0}{Characteristic impedance}
}
\description{
Conversion between the reflection coefficient (gamma), S11, impedance, and
mismatch gain:
}
\details{
\strong{Fundamental conversions:}\describe{
\item{\code{gamma_to_mismatch()}}{Converts the input reflection coefficient to
mismatch gain.} \item{\code{gamma_to_s11()}}{Converts the input reflection
coefficient to return \emph{gain}.} \item{\code{gamma_to_z()}}{Converts the input
reflection coefficient to impedance.  If the input is a complex number,
complex impedance will be output, and if the input is a scalar or magnitude,
only a scalar impedance is output.} \item{\code{mismatch_to_gamma()}}{Converts the
input mismatch gain to the reflection coefficient (magnitude only).}
\item{\code{s11_to_gamma()}}{Converts the input return \emph{gain} to the
reflection coefficient (magnitude only).} \item{\code{z_to_gamma()}}{Converts the
input impedance to the reflection coefficient.  If the input is a complex
number, the complex reflection coefficient will be output, and if the input
is a scalar or magnitude, only the reflection coefficient magnitude will be
output.} } \strong{Additional conversions:}\describe{
\item{\code{mismatch_to_s11()}}{Converts the input mismatch gain to return gain.}
\item{\code{s11_to_mismatch()}}{Converts the input return gain to mismatch gain.}
\item{\code{z_to_s11()}}{Converts the input impedance to return gain.}
\item{\code{z_to_mismatch()}}{Converts the input impedance to mismatch gain.} }
}
\examples{
# S11 --> reflection coefficient
s11_to_gamma(-3)

# Reflection coefficient --> S11
gamma_to_s11(0.5)

# Reflection coefficient --> impedance
gamma_to_z(0.5 + 0.5i, z0 = 50)
}
