% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mean.R
\name{linear_mean}
\alias{linear_mean}
\title{Linear Mean}
\usage{
linear_mean(x, scalar = 10, ...)
}
\arguments{
\item{x}{A vector of numeric values}

\item{scalar}{Scalar to be used in \code{as_decibel} and \code{as_linear} functions for conversion}

\item{...}{Addional methods passed to the \code{mean} function}
}
\description{
Converts a numeric vector from decibel format to linear format, calculates the mean value, and converts back to decibel format.
}
\examples{
linear_mean(-88:-90)
}
