% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_conversion.R
\name{network_conversion}
\alias{network_conversion}
\alias{abcd_to_h}
\alias{abcd_to_s}
\alias{abcd_to_t}
\alias{abcd_to_y}
\alias{abcd_to_z}
\alias{h_to_abcd}
\alias{h_to_s}
\alias{h_to_t}
\alias{h_to_y}
\alias{h_to_z}
\alias{s_to_abcd}
\alias{s_to_h}
\alias{s_to_t}
\alias{s_to_y}
\alias{s_to_z}
\alias{t_to_abcd}
\alias{t_to_s}
\alias{t_to_y}
\alias{t_to_z}
\alias{y_to_abcd}
\alias{y_to_h}
\alias{y_to_s}
\alias{y_to_t}
\alias{y_to_z}
\alias{z_to_abcd}
\alias{z_to_h}
\alias{z_to_s}
\alias{z_to_t}
\alias{z_to_y}
\title{Electrical Network Transformation}
\usage{
abcd_to_h(abcd, z0 = c(50, 50))

abcd_to_s(abcd, z0 = c(50, 50))

abcd_to_t(abcd, z0 = c(50, 50))

abcd_to_y(abcd, z0 = c(50, 50))

abcd_to_z(abcd, z0 = c(50, 50))

h_to_abcd(h, z0 = c(50, 50))

h_to_s(h, z0 = c(50, 50))

h_to_t(h, z0 = c(50, 50))

h_to_y(h, z0 = c(50, 50))

h_to_z(h, z0 = c(50, 50))

s_to_abcd(s, z0 = c(50, 50))

s_to_h(s, z0 = c(50, 50))

s_to_t(s, z0 = c(50, 50))

s_to_y(s, z0 = c(50, 50))

s_to_z(s, z0 = c(50, 50))

t_to_abcd(t, z0 = c(50, 50))

t_to_s(t, z0 = c(50, 50))

t_to_y(t, z0 = c(50, 50))

t_to_z(t, z0 = c(50, 50))

y_to_abcd(y, z0 = c(50, 50))

y_to_h(y, z0 = c(50, 50))

y_to_s(y, z0 = c(50, 50))

y_to_t(y, z0 = c(50, 50))

y_to_z(y, z0 = c(50, 50))

z_to_abcd(z, z0 = c(50, 50))

z_to_h(z, z0 = c(50, 50))

z_to_s(z, z0 = c(50, 50))

z_to_t(z, z0 = c(50, 50))

z_to_y(z, z0 = c(50, 50))
}
\arguments{
\item{abcd}{ABCD-Parameters}

\item{z0}{Characteristic Impedance}

\item{h}{h-Parameters}

\item{s}{S-Parameters}

\item{t}{t-Parameters}

\item{y}{Y-Parameters}

\item{z}{Z-Parameters}
}
\value{
Converted electrical parameters in a 2x2 matrix format
}
\description{
Converts electrical networks parameter matrices into equivalent types.
}
\section{Parameter format}{

\describe{
\item{2x2 Matrix}{\tabular{rr}{
S11 \tab S12\cr
S21 \tab S22
}}
\item{Vector of length 4}{
S11 S21 S12 S22
}}
}

\examples{
s_to_abcd(c(0.1, 0.5, 0.5, 0.1))
s_to_abcd(matrix(c(0.1, 0.5, 0.5, 0.1), ncol = 2))
s_to_abcd(matrix(c(0.1, 0.5, 0.5, 0.1), ncol = 2), z0 = c(50,100))
}
