% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_conversion.R
\name{network_gain}
\alias{network_gain}
\alias{available_gain}
\alias{power_gain}
\alias{transducer_gain}
\title{Electrical Network Gain}
\usage{
available_gain(S, gamma_source)

power_gain(S, gamma_load)

transducer_gain(S, gamma_source, gamma_load)
}
\arguments{
\item{S}{Network S-parameters,}

\item{gamma_source}{Source reflection coefficient}

\item{gamma_load}{Load reflection coefficient}
}
\value{
Gain value
}
\description{
Calculates various gain metrics for a two-port network
}
\examples{
S <- matrix(c(complex(modulus = 0.38, argument = -158 / 180 * pi),
              complex(modulus = 0.11, argument = 54 / 180 * pi),
              complex(modulus = 3.50, argument = 80 / 180 * pi),
              complex(modulus = 0.40, argument = -43 / 180 * pi)), nrow = 2, byrow = TRUE)

gamma_load <- z_to_gamma(40)
gamma_source <- z_to_gamma(25)

available_gain(S, gamma_source)
power_gain(S, gamma_load)
transducer_gain(S, gamma_source, gamma_load)

}
