% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_conversion.R
\name{network_input_output_gamma}
\alias{network_input_output_gamma}
\alias{gamma_in}
\alias{gamma_out}
\title{Input/Output Reflection Coefficient}
\usage{
gamma_in(S, gamma_load)

gamma_out(S, gamma_source)
}
\arguments{
\item{S}{Network S-parameters,}

\item{gamma_load}{Load reflection coefficient}

\item{gamma_source}{Source reflection coefficient}
}
\value{
Gain value
}
\description{
Calculates the input and output reflection coefficients of a two-port network
}
\examples{
S <- matrix(c(complex(modulus = 0.38, argument = -158 / 180 * pi),
              complex(modulus = 0.11, argument = 54 / 180 * pi),
              complex(modulus = 3.50, argument = 80 / 180 * pi),
              complex(modulus = 0.40, argument = -43 / 180 * pi)), nrow = 2, byrow = TRUE)
              
gamma_source <- z_to_gamma(25)
gamma_load <- z_to_gamma(40)

gamma_in(S, gamma_load)
gamma_out(S, gamma_source)

}
