% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{numeric_format}
\alias{numeric_format}
\alias{as_decibel}
\alias{as_linear}
\title{Number Conversion}
\usage{
as_decibel(..., scalar = 20)

as_linear(..., scalar = 20)
}
\arguments{
\item{...}{Number(s) to convert}

\item{scalar}{Decibel conversion scalar (i.e. 10 for power, 20 for
voltage/current)}
}
\description{
Conversions between linear and decibel number formats
}
\examples{
as_decibel(0.45)
as_decibel(0.45, scalar = 20)
as_decibel(c(0.1,0.5,0.9))
as_linear(-6)
as_linear(-6, scalar = 20)
as_linear(c(-12,-6,-3))
gamma_to_z
}
