% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_conversion.R
\name{param_coercion}
\alias{param_coercion}
\alias{coerce_params}
\alias{coerce_z0}
\title{Coerce Electrical Network Parameters Inputs}
\usage{
coerce_params(params)

coerce_z0(z0)
}
\arguments{
\item{params}{Input network parameters.  Must be one of:
\describe{\item{2x2 Matrix}{\tabular{rr}{ S11 \tab S12 \cr S21 \tab S22}}
\item{Vector of length 4}{S11 S21 S12 S22}
}}

\item{z0}{Network parameter characteristic impedance.  Must be one of:
\describe{
\item{\code{NA}, \code{NULL}}{Defaults to 50ohm input and output}
\item{Single numeric}{Same input and output characteristic impedance}
\item{Vector of length 2}{Input and output impedance, respectively}}}
}
\value{
Input coerced into a 2x2 matrix (params) or a vector of length 2 (z0) to be used in the conversions
}
\description{
Checks and converts the input parameters to the electrical network parameter conversion functions
}
\examples{
rftk:::coerce_params(c(0.1, 0.5, 0.5, 0.1))
rftk:::coerce_params(matrix(c(0.1, 0.5, 0.5, 0.1), ncol = 2))

rftk:::coerce_z0(50)
rftk:::coerce_z0(c(50,100))
}
\keyword{internal}
