% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactance.R
\name{reactance_conversion}
\alias{reactance_conversion}
\alias{c_to_x}
\alias{l_to_x}
\alias{x_to_c}
\alias{x_to_l}
\title{Reactance Conversion}
\usage{
c_to_x(c, freq)

l_to_x(l, freq)

x_to_c(x, freq)

x_to_l(x, freq)
}
\arguments{
\item{c}{Capacitance (F)}

\item{freq}{Frequency (Hz)}

\item{l}{Inductance (H)}

\item{x}{Reactance (jOhm)}
}
\value{
Converted data in SI units.
}
\description{
Converts between capacitance, inductance, and reactance
}
\examples{
c_to_x(1:10 * 1e-12, 2e9)
l_to_x(1:10 * 1e-9, 2e9)
x_to_c(-1:-10, 2e9) * 1e12
x_to_l(10:20, 2e9) * 1e9

}
