% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rftk.R
\docType{package}
\name{rftk-package}
\alias{rftk}
\alias{rftk-package}
\title{rftk:  RF Toolkit}
\description{
R functions for working with RF/wireless data.
}
\author{
\strong{Maintainer}: Aaron Anderson \email{aaron.m.anderson.3@gmail.com}

}
