% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{rftk_example}
\alias{rftk_example}
\title{Get path to rftk example}
\usage{
rftk_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
rftk comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
rftk_example()
rftk_example("dipole.s1p")
}
