% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_conversion.R
\name{series_shunt_impedance}
\alias{series_shunt_impedance}
\alias{abcd_to_series_z}
\alias{abcd_to_shunt_z}
\title{Series/Shunt Impedance}
\usage{
abcd_to_series_z(abcd)

abcd_to_shunt_z(abcd)
}
\arguments{
\item{abcd}{Network ABCD-parameters,}
}
\value{
Complex impedance
}
\description{
Calculates the series or shunt impedance based on ABCD-parameters
}
\examples{
abcd <- matrix(c(1,
                 complex(modulus = 50, argument = 0),
                 complex(modulus = 200, argument = pi / 2),
                 1), nrow = 2, byrow = TRUE)
              
z_series <- abcd_to_series_z(abcd)
z_shunt <- abcd_to_shunt_z(abcd)

}
